retriving data from json working
I need to add the images to the json so I can load them. I should also implement the onclick function, where I pass the Id and in the monster information I load it depending on a parameter passed that will tell me if it's mhw, mhr or mh4u. mby i could even split the id to see if the first part is "mhworld", "mhrise" or "mh4u"